back/requirements.txt
absl-py==2.1.0
asgiref==3.7.2
astunparse==1.6.3
Brotli==1.1.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
chex==0.1.85
click==8.1.7
colorama==0.4.6
dj-database-url==2.1.0
Django==4.0.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
dnspython==2.6.0
etils==1.7.0
filelock==3.13.1
flatbuffers==23.5.26
flax==0.8.1
fsspec==2024.2.0
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.20.3
idna==3.6
importlib-resources==6.1.1
jax==0.4.24
jaxlib==0.4.23
Jinja2==3.1.3
keras==2.15.0
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.7
mysqlclient==2.2.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.9
orbax-checkpoint==0.5.3
packaging==23.2
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
PyJWT==1.7.1
pymongo==3.12.3
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
safetensors==0.4.2
scipy==1.12.0
six==1.16.0
sqlparse==0.2.4
sympy==1.12
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorstore==0.1.53
termcolor==2.4.0
tokenizers==0.15.2
toolz==0.12.1
torch==2.2.1
tqdm==4.66.2
transformers==4.38.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
Werkzeug==3.0.1
whitenoise==6.6.0
wrapt==1.14.1
zipp==3.17.0
